backend/requirements.txt
fastapi
uvicorn
websockets
vaderSentiment
pydantic
python-multipart
requests
python-dotenv
tweepy
ntscraper
pandas
